@@ -11,7 +11,7 @@
     "delete": "node script/delete.js",
     "start": "npm run check && cross-env NODE_ENV=development webpack serve --hot --host 0.0.0.0 --config=./webpack.renderer.config.js --mode development",
     "prod": "npm run check && cross-env NODE_ENV=production webpack --mode production --config webpack.renderer.config.js && electron --noDevServer .",
-    "compile": "bytenode -c dist/**/*.js --electron && node script/compile.js",
+    "compile": "bytenode -c dist/main/*.js --electron && node script/compile.js",
     "build_renderer": "cross-env NODE_ENV=production webpack --config webpack.renderer.config.js --mode production",
     "build_main": "cross-env NODE_ENV=production webpack --config webpack.main.config.js --mode production",
     "build": "npm run delete && npm run build_renderer && npm run build_main && npm run compile",
@@ -28,8 +28,7 @@
     "files": [
       "dist/main/boot.js",
       "dist/main/main.jsc",
-      "dist/renderer/index.html",
-      "dist/renderer/main.jsc",
+      "dist/renderer/**/*",
       "node_modules/**/*",
       "!node_modules/**/*/{README,readme,LICENSE,license}",
       "!node_modules/**/*/*.md",
@@ -67,12 +66,12 @@
   },
   "devDependencies": {
     "electron": "30.4.0",
-    "electron-builder": "^25.1.8",
+    "electron-builder": "26.0.6",
     "electron-debug": "3.2.0",
     "electron-devtools-installer": "^3.2.0",
     "electron-reload": "1.5.0",
     "terser-webpack-plugin": "^5.3.11",
-    "webpack": "^5.75.0",
+    "webpack": "^5.98.0",
     "webpack-bundle-analyzer": "^4.7.0",
     "webpack-cli": "^5.0.1",
     "webpack-dev-server": "^4.11.1",
@@ -86,7 +85,7 @@
     "babel-loader": "^8.2.5",
     "bindings": "^1.5.0",
     "bufferutil": "^4.0.9",
-    "bytenode": "^1.5.7",
+    "bytenode": "1.5.6",
     "concurrently": "8.0.1",
     "cross-env": "7.0.3",
     "css-loader": "^6.7.1",
